e-lib/backend/requirements.txt
bibtexparser==1.4.0
blinker==1.9.0
cffi==2.0.0
click==8.3.0
cryptography==46.0.2
dnspython==2.8.0
email-validator==2.0.0
Flask==3.1.2
Flask-Cors==4.0.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pycparser==2.23
PyJWT==2.10.1
pymongo==4.5.0
pyparsing==3.2.5
python-dotenv==1.0.0
python-multipart==0.0.6
typing_extensions==4.15.0
Werkzeug==3.1.3
